Bug Fix
Cloud bases in the first radar range gate were not detected with this method, as it looked for groups of same values and used the number of values in each group to generate the index of the cloud bases and tops. Adding a first empty (no cloud) row to the mask makes sure that clouds which start at the first radar range gate are detected. However, it also increases the first discovered cloud base by one index, thus this needs to be accounted for.