Bug fix
Remove assertion statement to check if the lower range gat is lower than the upper range gate. If the range gates are the same, this also throws an error as happened for the 15. February.
It is also an unnecessary check as the if statement for the minimal vertical extent would filter those values as well.